Fully functional

+++ unpack_stems.sh
@@ -5,26 +5,41 @@
 # named after the stem filenames.
 #
 
-for stem in *.mp4
+sourceDir=~/Downloads/
+targetDir=~/Music/Ableton/User\ Library/MyInstuments/Stems/
+
+for stem in $sourceDir*.mp4
 do
-  name=${stem%\.*}
-  echo $name
+  if [[ $stem == *.stem.mp4 ]]
+  then
+    name=$(basename $stem)
+    name=${name%\.*}
+    dirName=$targetDir"$name"  
+    if [ ! -d "$dirName" ]; then
+      echo "Extracting $name..."
+      echo $dirName
 
-  # Create a directory
-  dirName=~/Music/Ableton/User\ Library/MyInstuments/Stems/"$name"
-  rm -R "$dirName"
-  mkdir "$dirName"
+      # Create a directory
+      dirName=$targetDir"$name"
+      rm -R "$dirName"
+      mkdir "$dirName"
 
-  # # Album cover
-  coverFilename="$dirName"/"$name"_cover.jpg
-  echo "  - Extracting album cover to $coverFilename"
-  ffmpeg -i "$stem" -an -vcodec copy "$coverFilename"
+      # # Album cover
+      coverFilename="$dirName"/"$name"_cover.jpg
+      echo "  - Extracting album cover to $coverFilename"
+      ffmpeg -i "$stem" -an -vcodec copy "$coverFilename"
 
-  # Audio tracks
-  for trackNo in 1 2 3 4 5 6 7
-  do
-    trackFilename="$dirName"/"$name"_"$trackNo".wav
-    echo "  - Extracting audio track $trackNo to $trackFilename"
-    ffmpeg -i $stem -map 0:$trackNo -vn "$trackFilename"
-  done
+      # Audio tracks
+      for trackNo in 1 2 3 4 5 6 7
+      do
+        trackFilename="$dirName"/"$name"_"$trackNo".wav
+        echo "  - Extracting audio track $trackNo to $trackFilename"
+        ffmpeg -i $stem -map 0:$trackNo -vn "$trackFilename"
+      done
+    else
+      echo "Already extracted $name"
+    fi
+  fi
+
 done
+
